Fix broken links on the 'tags' page
Some 'tags' links are broken, because the tag text itself may contain
unexpected characters (whitespace, periods, etc).  Scrub the link targets to
match to the actual file names on disk.

There's probably a better way to solve this.  Perhaps a dedicated filter
for scrubbing tags; or possibly smarter logic in generate_tags() in the Pelican
engine itself; but this works well enough for now.